Add options in run config to keep volumes and don't upload data. Adjusted shutdown scripts accordingly. Add default run config file and script to intialize run config

+++ server-setup/fhirbase/remove.sh
@@ -1,5 +1,28 @@
 #!/usr/bin/bash
 
+keep_volumes=false
+help=false
+
+while getopts vh flag
+do
+    case "${flag}" in
+        v) keep_volumes=true;;
+        h) help=true;;
+    esac
+done
+
+if [ "$help" = true ]; then
+  echo "Usage: bash remove.sh [flags]"
+  echo "Flags:"
+  echo "  -v: Don't remove volumes after container removal"
+  echo "  -h: Show this help message"
+  exit 0
+fi
+
 echo "Removing fhirbase FHIR server"
-docker compose --project-name feasibility-query-performance-fhirbase down --volumes
+if [ "$keep_volumes" = true ]; then
+  docker compose --project-name feasibility-query-performance-fhirbase down
+else
+  docker compose --project-name feasibility-query-performance-fhirbase down --volumes
+fi
 echo "Done"
